Fizz Buzz

+++ 0x04-more_functions_nested_loops/9-fizz_buzz.c
@@ -9,7 +9,7 @@
  */
 int main(void)
 {
-	unsigned long Q;
+	int Q;
 
 	for (Q = 1; Q <= 100; Q++)
 	{
@@ -26,7 +26,7 @@ int main(void)
 			printf("Buzz ");
 		}
 		else
-			printf("%lu ", Q);
+			printf("%i ", Q);
 	}
 	printf("\n");
 	return (0);
